Inventory routing confirmations

+++ src/app/inventory/edit/asset-edit.component.ts
@@ -4,6 +4,7 @@ import { Asset } from 'src/app/models/asset';
 import { InventoryService } from '../inventory.service';
 import { AuthService } from 'src/app/services/auth.service';
 import { NotificationService } from 'src/app/services/notification.service';
+import { cloneDeep } from 'lodash';
 
 @Component({
     templateUrl: 'asset-edit.component.html'
@@ -12,6 +13,8 @@ export class AssetEditComponent implements OnInit {
     public isEdit: boolean = false;
     public loading: boolean = false;
     public asset: Asset = new Asset();
+
+    private oAsset: Asset = new Asset();
 
 
     constructor(
@@ -31,6 +34,7 @@ export class AssetEditComponent implements OnInit {
                     this.inventoryService.getAsset(params.get('id')).subscribe(
                         (asset: Asset) => {
                             this.asset = asset;
+                            this.oAsset = cloneDeep(asset);
                             this.loading = false;
                         },
                         (error) => {
@@ -40,6 +44,7 @@ export class AssetEditComponent implements OnInit {
                     )
                 } else {
                     this.asset = new Asset();
+                    this.oAsset = new Asset();
                 }
             }
         );
@@ -58,6 +63,7 @@ export class AssetEditComponent implements OnInit {
                             this.router.navigateByUrl('inventory/' + asset._id);
                         } else {
                             this.asset = asset;
+                            this.oAsset = cloneDeep(asset);
                             this.loading = false;
                         }
                     });
@@ -70,25 +76,57 @@ export class AssetEditComponent implements OnInit {
         );
     }
 
-    
+    public openNew(): void {
+        if(JSON.stringify(this.asset) == JSON.stringify(this.oAsset)){
+            this.router.navigateByUrl('inventory/new');
+        } else {
+            this.notification.confirm('There are unsaved changes. Are you sure you want to leave?').subscribe(
+                (response: boolean) => {
+                    if(response){
+                        this.router.navigateByUrl('inventory/new');
+                    }
+                }
+            );
+        }
+    }
 
     public delete(): void {
-        this.loading = true;
-        this.inventoryService.deleteAsset(this.asset._id).subscribe(
-            (asset: Asset) => {
-                if(asset){
-                    this.router.navigateByUrl('inventory');
-                } else {
-                    this.asset = asset;
-                    this.notification.error('Error while deleting asset');
+        this.notification.confirm('Are you sure you want to delete this asset?').subscribe(
+            (response: boolean) => {
+                if(response){
+                    this.loading = true;
+                    this.inventoryService.deleteAsset(this.asset._id).subscribe(
+                        (asset: Asset) => {
+                            if(asset){
+                                this.router.navigateByUrl('inventory');
+                            } else {
+                                this.asset = asset;
+                                this.notification.error('Error while deleting asset');
+                            }
+                            this.loading = false;
+                        }, 
+                        (error) => {
+                            console.log(error);
+                            this.loading = false;
+                            this.notification.error('Error while deleting asset');
+                        }
+                    );
                 }
-                this.loading = false;
-            }, 
-            (error) => {
-                console.log(error);
-                this.loading = false;
-                this.notification.error('Error while deleting asset');
             }
         );
+    }
+
+    public close(): void {
+        if(JSON.stringify(this.asset) == JSON.stringify(this.oAsset)){
+            this.router.navigateByUrl('inventory');
+        } else {
+            this.notification.confirm('There are unsaved changes. Are you sure you want to leave?').subscribe(
+                (response: boolean) => {
+                    if(response){
+                        this.router.navigateByUrl('inventory');
+                    }
+                }
+            );
+        }
     }
 }
